requirements.txt
langchain
langchain-community
openai
chromadb
tiktoken
streamlit
pypdf
python-dotenv
